+++ common/kvstore/src/main/java/org/apache/spark/util/kvstore/InMemoryStore.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.spark.kvstore;
+package org.apache.spark.util.kvstore;
 
 import java.util.ArrayList;
 import java.util.Collection;
