+++ common/kvstore/src/test/java/org/apache/spark/util/kvstore/DBIteratorSuite.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.spark.kvstore;
+package org.apache.spark.util.kvstore;
 
 import java.util.Arrays;
 import java.util.ArrayList;
